requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
convertdate==2.4.0
cryptography==44.0.0
cssselect2==0.7.0
defusedxml==0.7.1
Django
django-auditlog==3.0.0
django-jalali==7.2.0
fonttools==4.55.3
fpdf2==2.8.2
jalali_core==1.0.0
jdatetime==5.0.0
lxml==5.3.0
Markdown==3.7
oscrypto==1.3.0
persiantools==5.1.0
pillow==11.0.0
psycopg2==2.9.10
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyMeeus==0.5.12
pypdf==5.1.0
pyphen==0.17.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
six==1.17.0
sqlparse==0.5.2
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tzlocal==5.2
uritools==4.0.3
urllib3==2.3.0
weasyprint==63.1
webencodings==0.5.1
xhtml2pdf==0.2.16
zopfli==0.2.3.post1
dj-database-url
python-dotenv